backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Celery
celery==5.3.6
flower==2.0.1
vine==5.1.0      # 固定版本，防止新版本移除 promise 接口导致 celery 启动失败
amqp==5.2.0      # celery 消息队列依赖，固定版本避免与 vine 不兼容

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Agent Libraries
openai==1.12.0
anthropic==0.18.1
langchain>=0.3.0              # LangChain 核心框架（0.3+ 支持 Tool Calling）
langchain-openai>=0.2.0       # OpenAI / 兼容 API 接入
langchain-google-genai>=2.0.0 # Google Gemini 接入
langchain-community>=0.3.0    # 社区工具包（FileManagementToolkit 等）
langchain-experimental>=0.3.0 # 实验性功能（PythonREPLTool、PandasDataframeAgent）
google-generativeai>=0.8.0    # Google Gemini 底层 SDK
pandas>=2.0.0                 # CSV 数据分析（CsvAgent 依赖）
duckduckgo-search>=6.0.0      # 免费网络搜索工具（无需 API Key）
pyyaml>=6.0                   # 读取 YAML 配置文件

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Utils
python-dotenv==1.0.1
email-validator==2.1.0
python-dateutil==2.8.2
pytz==2024.1

# Monitoring
prometheus-client==0.19.0

# Testing (dev)
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
